First iteration of the StatusBarItem
This 'StatusBarItem' improves the discoverability of the most important
interactions between the user and the extension.

This includes debug information, such as the version of the extension,
showing the output panel to find the logs more easily, but also to
restart all running Haskell Language Server binaries and the extension
itself.