Introduce 'automatic' mode for 'supportCabalFiles' option
Add option to automatically determine whether the Haskell Language
Server binary we are about to launch supports '.cabal' files.
If not, i.e. the HLS binary is older than '1.9.0.0', then we do not send
file notifications for '.cabal' files. Otherwise, we do send these
notifications.

The user can explicitly overwrite this option in order to avoid the
version check.